+++ simple_evals/common.py
@@ -7,7 +7,7 @@ import jinja2
 import numpy as np
 from tqdm import tqdm
 
-from simp_eval.types import EvalResult, Message, SamplerBase, SingleEvalResult
+from simple_evals.types import EvalResult, Message, SamplerBase, SingleEvalResult
 
 QUERY_TEMPLATE_MULTICHOICE = """
 Answer the following multiple choice question. The last line of your response should be of the following format: 'Answer: $LETTER' (without quotes) where LETTER is one of ABCD. Think step by step before answering.
