+++ simple_evals/tasks/drop_eval.py
@@ -15,9 +15,9 @@ import numpy as np
 from scipy.optimize import linear_sum_assignment
 import urllib.request
 
-from simp_eval import common
-from simp_eval.common import ANSWER_PATTERN, HTML_JINJA
-from simp_eval.types import Eval, EvalResult, SamplerBase, SingleEvalResult
+from simple_evals import common
+from simple_evals.common import ANSWER_PATTERN, HTML_JINJA
+from simple_evals.types import Eval, EvalResult, SamplerBase, SingleEvalResult
 
 """
 From here through _normalize_answer was originally copied from:
